+++ R/accounts.R
@@ -246,8 +246,8 @@ setAccountInfo <- function(name, token, secret,
     stop(stringParamErrorMessage("secret"))
 
   # create connect client
-  if (identical(server, cloudServerInfo()$name)) {
-    serverInfo <- cloudServerInfo()
+  if (identical(server, cloudServerInfo(server)$name)) {
+    serverInfo <- cloudServerInfo(server)
   } else {
     serverInfo <- shinyappsServerInfo()
   }
@@ -486,7 +486,8 @@ resolveAccount <- function(account, server = NULL) {
 }
 
 isCloudServer <- function(server) {
-  identical(server, "shinyapps.io") || identical(server, "rstudio.cloud")
+  identical(server, "shinyapps.io")||
+    identical(server, cloudServerInfo(server)$name)
 }
 
 isShinyappsServer <- function(server) {